refactor: GameTable pop을 giveCard로 변경
