+++ src/main/java/blackjack/domain/Playable.java
@@ -1,9 +1,12 @@
 package blackjack.domain;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.gamer.Participant;
 import java.util.List;
 
 public interface Playable {
+    void giveCard(Participant participant);
+
     Card pop();
 
     boolean isEmpty();
