+++ src/main/java/blackjack/domain/gamer/Participant.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
 
+import blackjack.domain.card.Cards;
 import blackjack.domain.card.Card;
 import java.util.List;
